requirements.txt
asgiref==3.8.1
Django==5.2.3
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
